增加 timer 线程，实现 fiber sleep。初步实现异步调用架构。

+++ fibjs/include/ClassInfo.h
@@ -131,7 +131,7 @@ protected:
 
         strError += *v8::String::Utf8Value(property);
         strError += "\' is read-only.";
-        ThrowException(v8::String::New(strError.c_str(), strError.length()));
+        ThrowException(v8::String::New(strError.c_str(), (int)strError.length()));
     }
 
 private:
